tools: fix Amplify build for valuation chart
Two issues from the previous PR (#1104) broke the production build:
- Missing @types/js-yaml caused a TS declaration error in valuationData.ts
- highcharts/highcharts-more side-effect import ran during SSR and tried
  to register against an uninitialized SeriesRegistry, breaking prerender
  for the United/Delta/Southwest tool pages.

Add @types/js-yaml as a devDependency and split ValuationChart into a
thin client wrapper that next/dynamic-loads the inner Highcharts impl
with ssr: false, so highcharts-more never loads during server render.
