+++ apps/web-next/package.json
@@ -36,6 +36,7 @@
     "@eslint/eslintrc": "^3.3.5",
     "@next/bundle-analyzer": "^16.1.6",
     "@types/canvas-confetti": "^1.9.0",
+    "@types/js-yaml": "^4.0.9",
     "@types/node": "^20",
     "@types/react": "^18",
     "@types/react-dom": "^18",
